virtio-gpu: Optimize flushing
First, queue both the TRANSFER_TO_HOST_2D and the RESOURCE_FLUSH
commands at the same time.  Second, do not wait for completion, instead
return immediately (without a result, but who would care about that
anyway?), and wait for the descriptors to be free before they are to be
used again.

Signed-off-by: Hanna Reitz <hanna@xanclic.moe>