virt-sound: Use sampling format of first track
Instead of fixing a sampling format, use the one of the first track that
is played.  Tracks following that one will only be played if they match.
So currently this means that the key press sound will only be played if
your music.ogg is in 8000 Hz mono, but maybe that's a good thing.

Signed-off-by: Hanna Reitz <hanna@xanclic.moe>