virtio-gpu/input: Pointing improvements
The virtio-gpu driver now has cursor support, and this is used to
display a PNG cursor.

However, this works really badly with a mouse input because of different
movement speeds in the host and the guest.  The cursor is in fact just
the host cursor (which simply gets a different sprite), so it is
repositioned automatically when the mouse is moved.  Then we try to move
it where we think it should be.  This results in bad jittering if we
don't agree with the host (and we don't).

So the virtio-input driver got tablet device support, which allows
absolute pointing device information.  With that, the jittering goes
away.  In fact, with that we can even skip updating the cursor
altogether because the host will do it automatically anyway.

Signed-off-by: Hanna Reitz <hanna@xanclic.moe>