Assets: Army counters up to 99; 2D region coll.
Army counters go up to 99 now (not sure if this is the most elegant
solution, but...).  Checking which region is hit by the mouse pointer is
now done with a map copy that is colored differently per region.  So
finding the region then is a simple lookup into that map copy.

Clicking on a country increases the army counter (no limit check yet, so
clicking enough gets you to the blue army).

Signed-off-by: Hanna Reitz <hanna@xanclic.moe>