Game logic: Same neutral army count for everyone
With the current numbers, the number of neutral armies to place after
the random auto-placement is uneven.  This gives the starting party an
unfair advantage (beyond the advantage they already have).  Increasing
the number so it is even is fair, since it does not really matter how
many neutral armies there are exactly.

Signed-off-by: Hanna Reitz <hanna@xanclic.moe>