Game logic: Add cards
This adds risk cards.  Support is incomplete; if you defeat an opponent,
get their cards and have more than five cards in hand, the game will
abort.  This case does not happen right now, as whenever you defeat an
opponent who has cards (i.e. not the neutral party), the game will also
end.

Implementing this should not be too hard, but with just two real
parties, I have no way of testing and there is no point in doing it.
This is why it is incomplete.

Signed-off-by: Hanna Reitz <hanna@xanclic.moe>