Assets: Change map
Surprise, surprise, it turns out any map of Middle Earth is fully
copyrighted.  Good thing I never added the map itself to the repo, I
suppose.

Anyway, the map is now replaced with something I made with
https://azgaar.github.io/Fantasy-Map-Generator .  The old map was kind
of low-detail because I wanted to evade copyright issues, but that is
out of the window now, so to make up for the loss in fun of playing on a
known map, the new map just looks much better.

Signed-off-by: Hanna Reitz <hanna@xanclic.moe>